+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - nodejs: wopi UserCanOnlyComment
 
 ## 1.15.0
+- nodejs: added AI Auto Fill Plugin
 - php-laravel: fix custom jwt header
 - formats for ds v9.3
 - user role